works on the way down, accounting is in place (sort of) way up to do next

+++ branches/tree-slurmd/src/common/slurm_protocol_defs.h
@@ -176,14 +176,14 @@ typedef struct slurm_protocol_header {
 	uint16_t flags;
 	slurm_msg_type_t msg_type;
 	uint32_t body_length;
-	uint32_t forward_cnt;
+	uint16_t forward_cnt;
 	slurm_addr *forward_addr;
 } header_t;
 
 typedef struct slurm_msg {
 	slurm_msg_type_t msg_type;
 	slurm_addr address;
-	uint32_t forward_cnt;
+	uint16_t forward_cnt;
 	slurm_addr *forward_addr;
 	slurm_fd conn_fd;
 	void *cred;
